+++ src/main/java/com/rcextract/minecord/DriverNotFoundException.java
@@ -1,7 +1,0 @@
-package com.rcextract.minecord;
-
-public class DriverNotFoundException extends Exception {
-
-	private static final long serialVersionUID = 7559113230833684270L;
-
-}