+++ src/main/java/com/rcextract/minecord/SQLConnectException.java
@@ -1,7 +1,0 @@
-package com.rcextract.minecord;
-
-public class SQLConnectException extends Exception {
-
-	private static final long serialVersionUID = 6251026970225795331L;
-
-}