+++ src/main/java/com/rcextract/minecord/Sendable.java
@@ -1,11 +1,11 @@
 package com.rcextract.minecord;
 
-import java.util.Map;
 import java.util.Set;
 
 import com.rcextract.minecord.getters.ChannelOptionsGetter;
+import com.rcextract.minecord.sql.DatabaseSerializable;
 
-public interface Sendable extends ChannelOptionsGetter {
+public interface Sendable extends ChannelOptionsGetter, DatabaseSerializable {
 	
 	public int getIdentifier();
 	public String getName();
@@ -15,5 +15,4 @@ public interface Sendable extends ChannelOptionsGetter {
 	public Set<Server> getServers();
 	public Channel getMain();
 	public void setMain(Channel channel);
-	public Map<String, Object> serialize();
 }