+++ src/main/java/com/rcextract/minecord/sql/DataLoadException.java
@@ -1,0 +1,14 @@
+package com.rcextract.minecord.sql;
+
+public class DataLoadException extends SQLObjectConnectorException {
+
+	private static final long serialVersionUID = 1315522023117907164L;
+
+	public DataLoadException() {
+		super();
+	}
+	
+	public DataLoadException(Throwable cause) {
+		super(cause);
+	}
+}