+++ src/main/java/com/rcextract/minecord/sql/DriverNotFoundException.java
@@ -1,0 +1,16 @@
+package com.rcextract.minecord.sql;
+
+import java.sql.SQLException;
+
+public class DriverNotFoundException extends SQLException {
+
+	private static final long serialVersionUID = 7559113230833684270L;
+
+	public DriverNotFoundException() {
+		super();
+	}
+	
+	public DriverNotFoundException(Throwable cause) {
+		super(cause);
+	}
+}