+++ src/main/java/com/rcextract/minecord/utils/Pair.java
@@ -7,8 +7,16 @@ public class Pair<K, V> implements Map.Entry<K, V> {
 	private K key;
 	private V value;
 
+	/**
+	 * Constructs a blank new <code>Pair</code>.
+	 */
 	public Pair() {}
 	
+	/**
+	 * Constructs a new <code>Pair</code> with default values.
+	 * @param key The key.
+	 * @param value The value.
+	 */
 	public Pair(K key, V value) {
 		this.key = key;
 		this.value = value;
@@ -24,6 +32,11 @@ public class Pair<K, V> implements Map.Entry<K, V> {
 		return value;
 	}
 
+	/**
+	 * Sets the key.
+	 * @param key The key.
+	 * @return The old key.
+	 */
 	public K setKey(K key) {
 		K oldkey = this.key;
 		this.key = key;